+++ pkgs/ffigen/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 20.1.0
+
+- Provide more convenience utils for building FFIgen configs:
+  `Declarations.excludeAll`, `Declarations.includeAll`,
+  `Declarations.includeSet`, `Declarations.includeAllMembers`,
+  `Declarations.includeMemberSet`, `Declarations.useOriginalName`,
+  `Declarations.renameWithMap`, `Declarations.useMemberOriginalName`, and
+  `Declarations.renameMemberWithMap`.
+
 ## 20.0.0
 
 - __Breaking change__: Completely rewrite the public Dart API for FFIgen.